Investment Case Study
The business objectives and goals of data analysis are pretty straightforward.

Business objective: The objective is to identify the best sectors, countries, and a suitable investment type for making investments. The overall strategy is to invest where others are investing, implying that the best sectors and countries are the ones where most investments are happening.

Goals of data analysis: Your goals are divided into three sub-goals:
Investment type analysis: Understanding investments in the venture, seed/angel, private equity categories, etc. so Spark Funds can decide which type is best suited for its strategy.

Country analysis: Understanding which countries have had the most investments in the past. These will be Spark Funds’ favourites as well.

Sector analysis: Understanding the distribution of investments across the eight main sectors. (Note that we are interested in the eight main sectors provided in the mapping file. The two files — companies and rounds2 — have numerous sub-sector names; hence, you will need to map each sub-sector to its main sector.)